+++ sway/commands/exec_always.c
@@ -8,6 +8,7 @@
 #include "sway/commands.h"
 #include "sway/config.h"
 #include "sway/tree/container.h"
+#include "sway/tree/root.h"
 #include "sway/tree/workspace.h"
 #include "log.h"
 #include "stringop.h"
