+++ sway/commands/urgent.c
@@ -4,7 +4,6 @@
 #include "sway/tree/arrange.h"
 #include "sway/tree/container.h"
 #include "sway/tree/view.h"
-#include "sway/tree/layout.h"
 #include "util.h"
 
 struct cmd_results *cmd_urgent(int argc, char **argv) {
