+++ sway/ipc-server.c
@@ -31,6 +31,7 @@
 #include "sway/server.h"
 #include "sway/input/input-manager.h"
 #include "sway/input/seat.h"
+#include "sway/tree/root.h"
 #include "sway/tree/view.h"
 #include "list.h"
 #include "log.h"
